hub/serializers.py
from rest_framework import serializers
from hub.models import Node, Task, TaskAssignment

class NodeSerializer(serializers.ModelSerializer):
    class Meta:
        model = Node
        fields = '__all__'

class TaskSerializer(serializers.ModelSerializer):
    class Meta:
        model = Task
        fields = '__all__'
